+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Implement cost center autocomplete, lazyload in organization autocomplete and maxHeight prop
+
 ## [0.3.2] - 2024-08-22
 
 ### Fixed
 - shows more options in the cost centers dropdown
+
 
 ### Removed
 - [ENGINEERS-1247] - Disable cypress tests in PR level